#1 Initial working configHandle

+++ g2config/main.go
@@ -31,16 +31,16 @@ const MessageIdFormat = "senzing-6014%04d"
 // ----------------------------------------------------------------------------
 
 type G2config interface {
-	AddDataSource(ctx context.Context, configHandle int64, inputJson string) (string, error)
+	AddDataSource(ctx context.Context, configHandle uintptr, inputJson string) (string, error)
 	ClearLastException(ctx context.Context) error
 	Close(ctx context.Context, configHandle uintptr) error
 	Create(ctx context.Context) (uintptr, error)
-	DeleteDataSource(ctx context.Context, configHandle int64, inputJson string) error
+	DeleteDataSource(ctx context.Context, configHandle uintptr, inputJson string) error
 	Destroy(ctx context.Context) error
 	GetLastException(ctx context.Context) (string, error)
 	GetLastExceptionCode(ctx context.Context) (int, error)
 	Init(ctx context.Context, moduleName string, iniParams string, verboseLogging int) error
-	ListDataSources(ctx context.Context, configHandle int64) (string, error)
+	ListDataSources(ctx context.Context, configHandle uintptr) (string, error)
 	Load(ctx context.Context, jsonConfig string) (string, error)
-	Save(ctx context.Context, configHandle int64) (string, error)
+	Save(ctx context.Context, configHandle uintptr) (string, error)
 }
